third_party/blink/web_tests/platform/generic/fast/dom/Element/class-list-update-expected.txt

Tests that modifying the class attribute updates the classList on non HTMLElements

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS htmlElement.classList.length is 1
PASS htmlElement.classList.length is 0
PASS svgElement.classList.length is 1
PASS svgElement.classList.length is 0
PASS xmlElement.classList.length is 1
PASS xmlElement.classList.length is 0
PASS successfullyParsed is true

TEST COMPLETE
